components/components-mongodb/components-mongodb-definition/src/main/java/org/talend/components/mongodb/tmongodbrow/TMongoDBRowDefinition.java
// ============================================================================
//
// Copyright (C) 2006-2017 Talend Inc. - www.talend.com
//
// This source code is available under agreement available at
// %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
//
// You should have received a copy of the agreement
// along with this program; if not, write to Talend SA
// 9 rue Pages 92150 Suresnes, France
//
// ============================================================================
package org.talend.components.mongodb.tmongodbrow;

import java.util.EnumSet;
import java.util.Set;

import org.talend.components.api.component.ConnectorTopology;
import org.talend.components.api.component.runtime.ExecutionEngine;
import org.talend.components.api.properties.ComponentProperties;
import org.talend.components.mongodb.common.MongoDBDefinition;
import org.talend.daikon.runtime.RuntimeInfo;

public class TMongoDBRowDefinition extends MongoDBDefinition {

    public static final String COMPONENT_NAME = "tMongoDBRow"; //$NON-NLS-1$

    public TMongoDBRowDefinition() {
        super(COMPONENT_NAME);
    }

    @Override
    public boolean isStartable() {
        return true;
    }

    @Override
    public String getPartitioning() {
        return NONE;
    }

    @Override
    public boolean isSchemaAutoPropagate() {
        return true;
    }

    @Override
    public Class<? extends ComponentProperties> getPropertyClass() {
        return TMongoDBRowProperties.class;
    }

    @Override
    public RuntimeInfo getRuntimeInfo(ExecutionEngine engine, ComponentProperties properties,
            ConnectorTopology connectorTopology) {
        assertEngineCompatibility(engine);
        assertConnectorTopologyCompatibility(connectorTopology);
        // TODO recheck when work on runtime part
        if (connectorTopology == ConnectorTopology.INCOMING) {
            return getCommonRuntimeInfo(ROW_SINK_CLASS);
        } else if (connectorTopology == ConnectorTopology.NONE) {
            return getCommonRuntimeInfo(ROW_SINK_OR_CLASS);
        } else {
            return null;
        }
    }

    @Override
    public Set<ConnectorTopology> getSupportedConnectorTopologies() {
        return EnumSet.of(ConnectorTopology.INCOMING, ConnectorTopology.NONE);
    }

}